genesis-core/src/config/anatomy.rs: Created anatomy.toml parser. Renamed the layer field to plural (layers) for architectural consistency (like inputs and neuron_types). Added population calculation tests.
genesis-core/src/config/io.rs: Created a unified parser for external io.toml projections. Also renamed the input field to inputs. Added a unit test.

Baker refactoring:
genesis-baker/src/parser/anatomy.rs and io.rs are now thin clients (re-exported from core).
Updated calls throughout baker (in the axon growth, validator, and neuron placement modules) (from anatomy.layer to anatomy.layers, from io.input to io.inputs).
The names of the validation functions in validator/checks.rs have been updated to match convention (e.g., check_layer_height_sum -> check_layer_heights).

Configuration templates:
The zones/_template directory has been created, which serves as the starting point for creating new zones.
A minimal working zone has been written:
anatomy.toml (1 kernel layer, 100% height, 2 neuron types).
blueprints.toml (2 defined types: excitatory and inhibitory).
io.toml (empty template with a commented example of an external connection).

Comprehensive reference files have been created:
_blank_anatomy.toml
_blank_blueprints.toml
_blank_io.toml. These files document each possible parameter in Russian, specifying the data types, value ranges, default values, and what exactly each parameter affects in the C++ (CUDA) or Rust engine logic.

+++ genesis-runtime/src/main.rs
@@ -156,7 +156,7 @@ async fn main() -> Result<()> {
             rest_potential:       nt.rest_potential,
             leak:                 nt.leak_rate,
             homeostasis_penalty:  nt.homeostasis_penalty,
-            homeostasis_decay:    nt.homeostasis_decay,
+            homeostasis_decay:    nt.homeostasis_decay as i32,
             gsop_potentiation:    nt.gsop_potentiation,
             gsop_depression:      nt.gsop_depression,
             refractory_period:    nt.refractory_period,
